Add scaffold - refactored scaffold to JokePresentation.kt - renamed "jokePresentation" package to "screens"
